requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
alabaster=0.7.10=py35hb692fe1_0
anaconda=custom=py35hd7b5ba2_0
anaconda-client=1.6.9=py35_0
anaconda-project=0.8.2=py35ha24014a_0
appnope=0.1.0=py35hd172556_0
appscript=1.0.1=py35hdfff33c_1
asn1crypto=0.24.0=py35_0
astroid=1.6.1=py35_0
astropy=2.0.3=py35h1de35cc_0
attrs=17.4.0=py35_0
babel=2.5.3=py35_0
backports=1.0=py35h2556c16_1
backports.shutil_get_terminal_size=1.0.0=py35h40fcb1f_2
beautifulsoup4=4.6.0=py35hb75f182_1
bitarray=0.8.1=py35h1de35cc_1
bkcharts=0.2=py35h18a45e2_0
blaze=0.11.3=py35h385aa17_0
bleach=2.1.2=py35_0
bokeh=0.12.13=py35h2f9c1c0_0
boto=2.48.0=py35h10f7326_1
bottleneck=1.2.1=py35h969083f_0
bzip2=1.0.6=hd86a083_4
ca-certificates=2018.03.07=0
certifi=2018.4.16=py35_0
cffi=1.11.4=py35h342bebf_0
chardet=3.0.4=py35h16a84c2_1
click=6.7=py35hcc65ea6_0
cloudpickle=0.5.2=py35_1
clyent=1.2.2=py35h63ae3d7_0
colorama=0.3.9=py35h1d66b2b_0
contextlib2=0.5.5=py35hc2c67b4_0
cryptography=2.1.4=py35hbbb57dc_0
curl=7.58.0=ha441bb4_0
cycler=0.10.0=py35hb89929e_0
cython=0.27.3=py35h36d65b6_0
cytoolz=0.9.0=py35h1de35cc_0
dask=0.16.1=py35_0
dask-core=0.16.1=py35_0
datashape=0.5.4=py35hd065018_0
dbus=1.12.2=h5243cc1_1
decorator=4.2.1=py35_0
distributed=1.20.2=py35_0
docutils=0.14=py35hb13dbd9_0
entrypoints=0.2.3=py35h837ec6f_2
et_xmlfile=1.0.1=py35h40eb147_0
expat=2.2.5=hb8e80ba_0
fastcache=1.0.2=py35h1de35cc_2
filelock=2.0.13=py35haecd4ea_0
flask=0.12.2=py35h7284a24_0
flask-cors=3.0.3=py35h7e589ad_0
freetype=2.8=h12048fb_1
get_terminal_size=1.0.0=h7520d66_0
gettext=0.19.8.1=h15daf44_3
gevent=1.2.2=py35h8412070_0
glib=2.53.6=h33f6a65_2
glob2=0.6=py35hdef87b6_0
gmp=6.1.2=hb37e062_1
gmpy2=2.0.8=py35h212fb8a_2
greenlet=0.4.12=py35h49b4c26_0
h5py=2.7.1=py35he1c7800_0
hdf5=1.10.1=ha036c08_1
heapdict=1.0.0=py35_2
html5lib=1.0.1=py35h2f9c1c0_0
icu=58.2=h4b95b61_1
idna=2.6=py35h01aacb0_1
imageio=2.2.0=py35h805106e_0
imagesize=0.7.1=py35h1ad78fc_0
intel-openmp=2018.0.0=h8158457_8
ipykernel=4.8.0=py35_0
ipython=6.2.1=py35h9470683_1
ipython_genutils=0.2.0=py35hf129286_0
ipywidgets=7.1.1=py35_0
isort=4.2.15=py35h0db285e_0
itsdangerous=0.24=py35hfbd69cd_1
jbig=2.1=h4d881f8_0
jdcal=1.3=py35h9028778_0
jedi=0.11.1=py35_0
jinja2=2.10=py35h6ff70ae_0
jpeg=9b=he5867d9_2
jsonschema=2.6.0=py35h2dd9e4b_0
jupyter=1.0.0=py35_4
jupyter_client=5.2.2=py35_0
jupyter_console=5.2.0=py35hd2aa692_1
jupyter_core=4.4.0=py35h4ad9194_0
jupyterlab=0.31.5=py35_0
jupyterlab_launcher=0.10.2=py35_0
lazy-object-proxy=1.3.1=py35h7293e74_0
libcurl=7.58.0=hf30b1f0_0
libcxx=4.0.1=h579ed51_0
libcxxabi=4.0.1=hebd6815_0
libedit=3.1=hb4e282d_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libpng=1.6.34=he12f830_0
libsodium=1.0.15=hd9e47c5_0
libssh2=1.8.0=h322a93b_4
libtiff=4.0.9=hcb84e12_1
libxml2=2.9.7=hab757c2_0
libxslt=1.1.32=hb819dd2_0
llvmlite=0.21.0=py35h3a66efe_0
locket=0.2.0=py35h58cf053_1
lxml=4.1.1=py35hef8c89e_1
lzo=2.10=h362108e_2
markupsafe=1.0=py35h9ba0a7f_1
matplotlib=2.1.2=py35h6d6146d_0
mccabe=0.6.1=py35h3f6a9a1_0
mistune=0.8.3=py35_0
mkl=2018.0.1=hfbd8650_4
mkl-service=1.1.2=py35h6a6947a_4
mpc=1.0.3=h7a72875_5
mpfr=3.1.5=h711e7fd_2
mpmath=1.0.0=py35he743aed_2
msgpack-python=0.5.1=py35h04f5b5a_0
multipledispatch=0.4.9=py35h2fbb584_0
nbconvert=5.3.1=py35h63fb950_0
nbformat=4.4.0=py35h41c2038_0
ncurses=6.0=hd04f020_2
networkx=2.1=py35_0
ninja=1.8.2=py35h04f5b5a_1
nltk=3.2.5=py35h87b897b_0
nose=1.3.7=py35h9ce1e3a_2
notebook=5.4.0=py35_0
numba=0.36.2=np114py35h86c4cab_0
numexpr=2.6.4=py35hcf51bc4_0
numpy=1.14.0=py35h8a80b8c_1
numpydoc=0.7.0=py35h296b98c_0
odo=0.5.1=py35h7f7a387_0
olefile=0.45.1=py35_0
openpyxl=2.4.10=py35_0
openssl=1.0.2o=h26aff7b_0
packaging=16.8=py35h2e4dc9b_0
pandas=0.22.0=py35h0a44026_0
pandoc=1.19.2.1=ha5e8f32_1
pandocfilters=1.4.2=py35hff87490_1
parso=0.1.1=py35hbda7c10_0
partd=0.3.8=py35h6fadee7_0
path.py=10.5=py35hcf34ba7_0
pathlib2=2.3.0=py35hba2ddec_0
patsy=0.5.0=py35_0
pcre=8.41=hfb6ab37_1
pep8=1.7.1=py35_0
pexpect=4.3.1=py35_0
pickleshare=0.7.4=py35h9517181_0
pillow=5.0.0=py35hfcce615_0
pip=9.0.1=py35h33ce766_4
pkginfo=1.4.1=py35h568fed1_0
pluggy=0.6.0=py35hf57b818_0
ply=3.10=py35hd0aaff5_0
prompt_toolkit=1.0.15=py35h93950c5_0
psutil=5.4.3=py35h1de35cc_0
ptyprocess=0.5.2=py35hfc37984_0
py=1.5.2=py35h4d4d87d_0
pycodestyle=2.3.1=py35he0976b1_0
pycosat=0.6.3=py35h745f8c1_0
pycparser=2.18=py35hab820b0_1
pycrypto=2.6.1=py35h1de35cc_7
pycurl=7.43.0.1=py35hdbc3d79_0
pyflakes=1.6.0=py35hc517269_0
pygments=2.2.0=py35h392a662_0
pylint=1.8.2=py35_0
pyodbc=4.0.22=py35h0a44026_0
pyopenssl=17.5.0=py35h4065cf8_0
pyparsing=2.2.0=py35h31fab04_0
pyqt=5.6.0=py35hbd126f6_6
pysocks=1.6.7=py35h3cfcbe1_1
pytables=3.4.2=py35hda701c8_2
pytest=3.3.2=py35_0
python=3.5.4=he720263_23
python-dateutil=2.6.1=py35h10515e0_1
python.app=2=py35he4d1c94_7
pytorch=0.4.0=py35_cuda0.0_cudnn0.0_1
pytz=2017.3=py35heeb7564_0
pywavelets=0.5.2=py35h9dc8fb8_0
pyyaml=3.12=py35hf8cec8a_1
pyzmq=16.0.3=py35hab9903b_0
qt=5.6.2=h9975529_14
qtawesome=0.4.4=py35h21e61ad_0
qtconsole=4.3.1=py35hd6d667b_0
qtpy=1.3.1=py35h743f03e_0
readline=7.0=hc1231fa_4
requests=2.18.4=py35h0d65e6b_1
rope=0.10.7=py35h27868a4_0
ruamel_yaml=0.15.35=py35h1de35cc_1
scikit-image=0.13.1=py35h1de35cc_1
scikit-learn=0.19.1=py35h2b554eb_0
scipy=1.0.0=py35h8b35106_0
seaborn=0.8.1=py35he0bbe96_0
send2trash=1.4.2=py35_0
setuptools=38.4.0=py35_0
simplegeneric=0.8.1=py35_2
singledispatch=3.4.0.3=py35h0acf360_0
sip=4.18.1=py35h79e1f92_2
six=1.11.0=py35h39a4c60_1
snowballstemmer=1.2.1=py35hbb7be01_0
sortedcollections=0.5.3=py35hefac527_0
sortedcontainers=1.5.9=py35_0
sphinx=1.6.6=py35_0
sphinxcontrib=1.0=py35h3eabf46_1
sphinxcontrib-websupport=1.0.1=py35hcb4ca16_1
spyder=3.2.6=py35_0
sqlalchemy=1.2.1=py35h1de35cc_0
sqlite=3.22.0=h3efe00b_0
statsmodels=0.8.0=py35ha7c9052_0
sympy=1.1.1=py35he478fab_0
tblib=1.3.2=py35h1b9c5fd_0
terminado=0.8.1=py35_1
testpath=0.3.1=py35hf8009f4_0
tk=8.6.7=h35a86e2_3
toolz=0.9.0=py35_0
torchvision=0.2.1=py35_1
tornado=4.5.3=py35_0
traitlets=4.3.2=py35hd3d1486_0
typing=3.6.2=py35ha402c5e_0
unicodecsv=0.14.1=py35h2154ad0_0
unixodbc=2.3.4=h4cb4dde_1
urllib3=1.22=py35he002d57_0
wcwidth=0.1.7=py35hdd0c235_0
webencodings=0.5.1=py35hcf8ebf9_1
werkzeug=0.14.1=py35_0
wheel=0.30.0=py35h5c0b906_1
widgetsnbextension=3.1.0=py35_0
wrapt=1.10.11=py35ha18cf31_0
xlrd=1.1.0=py35h892fcf7_1
xlsxwriter=1.0.2=py35h4d68a89_0
xlwings=0.11.5=py35_0
xlwt=1.2.0=py35hf70d7fa_0
xz=5.2.3=h727817e_4
yaml=0.1.7=hc338f04_2
zeromq=4.2.2=ha360ad0_2
zict=0.1.3=py35h1ae85d2_0
zlib=1.2.11=hf3cbc9b_2
